requirements.txt
flask~=3.1.2
python-dotenv~=1.1.1
requests~=2.32.5
psutil~=7.1.0